usersystemlib: split into files
